pyproject.toml
[project]
name = "langgraph-checkpoint-cassandra"
version = "0.1.0"
description = "LangGraph checkpointer for Cassandra"
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.6.9",
    "cassandra-driver",
]

[project.urls]
Repository = "https://github.com/johntrimble/langgraph-checkpoint-cassandra"

[build-system]
requires = ["uv_build>=0.8.24,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

[dependency-groups]
dev = [
    "cassandra-driver>=3.29.2",
    "cassandra-asyncio-driver>=0.1.4",
    "hypothesis>=6.140.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.8.6",
    "mypy>=1.18.2",
]

[tool.ruff]
lint.select = [
  "E",  # pycodestyle
  "F",  # Pyflakes
  "UP", # pyupgrade
  "B",  # flake8-bugbear
  "I",  # isort
]
lint.ignore = ["E501", "B008"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Respect magic trailing comma
skip-magic-trailing-comma = false
# Automatically detect line ending
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
pretty = true
namespace_packages = true
files = [
    "langgraph_checkpoint_cassandra",
    "tests",
]

[[tool.mypy.overrides]]
module = [
  "cassandra.*",
  "cassandra_asyncio.*",
  "cassandra_asyncio_driver.*",
  "aiofiles",
  "aiofiles.*",
  "msgpack",
  "msgpack.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
